Fix following splint warnings.
src/util.c: (in function fmt_loc)
src/util.c:313:50: Format argument 2 to fprintf (%d) expects int gets
uint16_t:
                      loc->first_line
  To allow arbitrary integral types to match any integral type, use
  +matchanyintegral.
   src/util.c:313:33: Corresponding format code
src/util.c:335:16: Operands of < have incompatible types (int, arbitrary
                      unsigned integral type): j < loc->first_column + 4
src/util.c:338:16: Operands of < have incompatible types (int, arbitrary
    unsigned integral type): j < last_col - loc->first_column + 1

+++ src/util.c
@@ -305,12 +305,12 @@ void fmt_loc(FILE *f, const struct loc *loc)
    size_t i = 0;
    bool many_lines;
    int last_col;
-   int j;
+   uint16_t j;
 
    if ((loc == NULL) || (loc->first_line == LINE_INVALID))
       return;
 
-   fprintf(f, "\tFile %s, Line %d\n", loc->file, loc->first_line);
+   fprintf(f, "\tFile %s, Line %u\n", loc->file, loc->first_line);
 
    if (loc->linebuf == NULL)
       return;
